% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calcola_generalized_factorial_ricor_log_old}
\alias{calcola_generalized_factorial_ricor_log_old}
\title{Calcola Fattoriale Generalizzato Ricorsivo Log}
\usage{
calcola_generalized_factorial_ricor_log_old(gamma, n)
}
\description{
Calcola Fattoriale Generalizzato Ricorsivo Log
}
