% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calcola_stirling_ricor_abs_old}
\alias{calcola_stirling_ricor_abs_old}
\title{Calcola Stirling Ricorsivo Abs}
\usage{
calcola_stirling_ricor_abs_old(gamma, n)
}
\description{
Calcola Stirling Ricorsivo Abs
}
