% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calcola_stirling_ricor_centr}
\alias{calcola_stirling_ricor_centr}
\title{Compute Central C number - Recursive formula Raf}
\usage{
calcola_stirling_ricor_centr(gamma, n)
}
\description{
Same as \code{\link{calcola_stirling_ricor}} but with r=0, i.e for central numbers only.
}
