% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calcola_stirling_ricor_log_centrali}
\alias{calcola_stirling_ricor_log_centrali}
\title{Compute log of absolute values of Central C number - Recursive formula for Raf}
\usage{
calcola_stirling_ricor_log_centrali(n, gamma)
}
\description{
Same as \code{\link{calcola_stirling_ricor_log}} but with r=0, i.e for central numbers only.
}
