% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calcola_stirling_ricor_log_old}
\alias{calcola_stirling_ricor_log_old}
\title{Calcola Stirling Ricorsivo Log}
\usage{
calcola_stirling_ricor_log_old(gamma, n)
}
\description{
Calcola Stirling Ricorsivo Log
}
