% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDFMM_export.R
\name{Compute_coclust_error}
\alias{Compute_coclust_error}
\title{Compute_coclust_error}
\usage{
Compute_coclust_error(real_partition, psm)
}
\arguments{
\item{real_partition}{\link{vector} of length \code{n}, the total number of points, containing the true cluster membership of each data point}

\item{psm}{\link{matrix} of size \code{n x n} containing the posterior similarity matrix, i.e each element represents the relative frequency of two data points being assigned to the same cluster}
}
\value{
\link{list} \code{coclust_err} and \code{coclust_err_star}
}
\description{
This function computes the coclustering error and the coclustering error star.
Being errors, low values are to be preferred and 1 is the maximum, worst, value.
See Bassetti,Casarin et al. 2020 for proper definition
}
