% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDFMM_export.R
\name{Compute_L1_dist}
\alias{Compute_L1_dist}
\title{Compute_L1_dist}
\usage{
Compute_L1_dist(Pred, p_mix, mu, sigma, grid)
}
\arguments{
\item{Pred}{\link{list} the predictive distributions of all levels evaluated  \code{grid}}

\item{mu}{\link{vector} of length \code{K} defining the means of the clusters.}

\item{grid}{\link{vector} a grid where the normal kernel is evaluated.}
}
\value{
\link{list} the first element, ??? , is a vector with the L1 error within each level.
The second element, ???, is the mean value across all levels.
}
\description{
This function computes the L1 distance between the true density and the predictive density within each level
and the mean value across all levels.
Low values are to be preferred.
}
\examples{
Compute_L1_dist(list(Pred_all[[1]][2,], Pred_all[[2]][2,]),mix_probs,mu,sd)
}
