% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDFMM_export.R
\name{ConditionalSampler}
\alias{ConditionalSampler}
\title{Conditional Sampler: function to run the GDFMM model. There is the possibility to fix
the partition, passing TRUE to FixPartition and specifying the
partion in the option. Default prior for P0 is an inverse gamma}
\usage{
ConditionalSampler(
  data,
  niter,
  burnin,
  thin,
  seed,
  P0.prior = "Normal-InvGamma",
  FixPartition = F,
  algorithm = "Neal2",
  option = NULL
)
}
\arguments{
\item{data}{\link{tibble}, the input data. This must be the return object of \code{\link{handle_input}}}

\item{niter}{\link{integer}, the number of iterations}

\item{burnin}{\link{integer}, the burnin period}

\item{thin}{\link{integer}, the thinning value}

\item{seed}{\link{integer}, the seed for GSL random engine (0 ==> random seed)}

\item{P0.prior}{\link{string} with the prior to be used as P0}

\item{FixPartition}{TRUE if we want to fix the partition}

\item{option}{\link{list} the output of \code{\link{set_options}} function}
}
\description{
Conditional Sampler: function to run the GDFMM model. There is the possibility to fix
the partition, passing TRUE to FixPartition and specifying the
partion in the option. Default prior for P0 is an inverse gamma
}
