% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDFMM_export.R
\name{data_mat2list}
\alias{data_mat2list}
\title{data_mat2list}
\usage{
data_mat2list(data)
}
\arguments{
\item{data}{\link{matrix} input data in long form}
}
\description{
function to transform the data matrix in long form in a list of vectors form.
}
