% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDFMM_export.R
\name{dmix}
\alias{dmix}
\title{dmix : density of a mixture of gaussian distributions}
\usage{
dmix(x, w_j, mu_vec, sigma_vec)
}
\arguments{
\item{x}{point or vector of values where I want to evaluate the density}

\item{mu_vec}{vector of mean for the gaussian distribution}

\item{sigma_vec}{vector of standard deviation for the gaussian distribution}

\item{w}{vector of weights of the}
}
\value{
named matrix with (Inf, estimate, Sup)
}
\description{
dmix : density of a mixture of gaussian distributions
}
