% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDFMM_export.R
\name{dnct}
\alias{dnct}
\title{Non Central Student t - Density}
\usage{
dnct(x, n0, mu0, gamma0)
}
\arguments{
\item{n0}{the degree of fredoom.}

\item{mu0}{the location parameter.}

\item{gamma0}{the scale parameter.}
}
\value{
scalar representing the evaluation of the density at x
}
\description{
Non Central Student t - Density
}
