% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDFMM_export.R
\name{empirical_bayes_normalinvgamma}
\alias{empirical_bayes_normalinvgamma}
\title{empirical_bayes_normalinvgamma}
\usage{
empirical_bayes_normalinvgamma(
  data = NULL,
  barmu = 0,
  varmu = 1,
  barsig2 = 10,
  varsig2 = 5,
  correction = 3
)
}
\arguments{
\item{data}{\link{matrix} input data.}

\item{barmu}{\link{scalar} desired value of the marginal mean of the normally distribuited parameter. Used only if \code{data} is \code{NULL}.}

\item{varmu}{\link{scalar} desired value of the marginal variance of the normally distribuited parameter}

\item{barsig2}{\link{scalar} desired value of the mean of the inverse-gamma distribuited parameter. Used only if \code{data} is \code{NULL}.}

\item{varsig2}{\link{scalar} desired value of the variance of the inverse-gamma distribuited parameter}

\item{correction}{\link{scalar} if \code{data} are provided, the expected value of \code{sigma} is set equal to \code{var(data)/correction}.}
}
\description{
function to set normal-inversegamma parameters. If \code{data} is not \code{NULL}, an empirical bayes procedure
is automatically applied by setting \code{barmu = mean(data)}, \code{barsig2 = var(as.vector(data))/3}
\code{varmu} and \code{varsig2} are not set using \code{data}
}
