% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDFMM_export.R
\name{example_GDFMM_sampler}
\alias{example_GDFMM_sampler}
\title{The R function}
\usage{
example_GDFMM_sampler(
  data,
  niter,
  burnin,
  thin,
  P0.prior = "Normal-InvGamma",
  option
)
}
\arguments{
\item{data}{input data.}

\item{niter}{number of iterations}

\item{burnin}{burnin period}

\item{thin}{thinning value}

\item{P0.prior}{string with the prior to be used as P0}

\item{option}{list with initial values, hyperparameters and other options}
}
\value{
what it returns
}
\description{
The R function
}
