% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDFMM_export.R
\name{GDFMM_marginal_sampler}
\alias{GDFMM_marginal_sampler}
\title{GDFMM Marginal Gibbs Sampler:}
\usage{
GDFMM_marginal_sampler(
  data,
  niter,
  burnin,
  thin,
  seed,
  P0.prior = "Normal-InvGamma",
  FixPartition = F,
  option = NULL
)
}
\arguments{
\item{data}{input data}

\item{niter}{number of iterations}

\item{burnin}{burnin period}

\item{thin}{thinning value}

\item{seed}{seed for GSL random engine (0 is for random seed)}

\item{P0.prior}{string with the prior to be used as P0}

\item{FixPartition}{TRUE if we want to fix the partition}

\item{option}{the output of \code{\link{set_options_marginal}} function}
}
\value{
results of Gibbs Sampler
}
\description{
function to run the GDFMM marginal Gibbs Sampler. There is the possibility to fix
the partition, passing TRUE to FixPartition and specifying the
partion in the option. Default prior for P0 is an inverse gamma
}
