% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GDFMM_marginal_sampler_c}
\alias{GDFMM_marginal_sampler_c}
\title{GDFMM sampler}
\usage{
GDFMM_marginal_sampler_c(
  dat,
  n_iter,
  burn_in,
  thin,
  seed,
  P0_prior_name,
  FixPart,
  option
)
}
\description{
GDFMM sampler
}
