% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDFMM_export.R
\name{GDFMM_sampler}
\alias{GDFMM_sampler}
\title{GDFMM Gibbs Sampler: function to run the GDFMM model. There is the possibility to fix
the partition, passing TRUE to FixPartition and specifying the
partion in the option. Default prior for P0 is an inverse gamma}
\usage{
GDFMM_sampler(
  data,
  niter,
  burnin,
  thin,
  seed,
  P0.prior = "Normal-InvGamma",
  FixPartition = F,
  option = NULL
)
}
\arguments{
\item{data}{input data}

\item{niter}{number of iterations}

\item{burnin}{burnin period}

\item{thin}{thinning value}

\item{seed}{seed for GSL random engine (0 ==> random seed)}

\item{P0.prior}{string with the prior to be used as P0}

\item{FixPartition}{TRUE if we want to fix the partition}

\item{option}{the output of \code{\link{set_options}} function}
}
\value{
results of Gibbs Sampler
}
\description{
GDFMM Gibbs Sampler: function to run the GDFMM model. There is the possibility to fix
the partition, passing TRUE to FixPartition and specifying the
partion in the option. Default prior for P0 is an inverse gamma
}
