% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GDFMM_sampler_c}
\alias{GDFMM_sampler_c}
\title{GDFMM sampler with M unknown}
\usage{
GDFMM_sampler_c(dat, n_iter, burn_in, thin, seed, P0_prior_name, option)
}
\description{
GDFMM sampler with M unknown
}
