% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDFMM_export.R
\name{GDFMM_sampler_M}
\alias{GDFMM_sampler_M}
\title{GDFMM Gibbs Sampler, M fixed}
\usage{
GDFMM_sampler_M(
  data,
  M,
  niter,
  burnin,
  thin,
  seed,
  P0.prior = "Normal-InvGamma",
  option
)
}
\arguments{
\item{data}{input data}

\item{M}{fixed number of components}

\item{niter}{number of iterations}

\item{burnin}{burnin period}

\item{thin}{thinning value}

\item{P0.prior}{string with the prior to be used as P0}

\item{option}{list with initial values, hyperparameters and other options}
}
\value{
results of Gibbs Sampler
}
\description{
GDFMM Gibbs Sampler, M fixed
}
