% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDFMM_export.R
\name{genera_mix_gas}
\alias{genera_mix_gas}
\title{genera_mix_gas}
\usage{
genera_mix_gas(
  n = 200,
  pro = c(0.5, 0.5),
  means = c(-1, 1),
  sds = sqrt(c(1, 1))
)
}
\arguments{
\item{n}{\link{integer} the number of points to be generated.}

\item{pro}{\link{vector} the mixture proportions. Its length must be equal to the one of \code{means} and \code{sds}.}

\item{means}{\link{vector} the centers of the mixture componentes. Its length must be equal to the one of \code{pro} and \code{sds}.}

\item{sds}{\link{vector} the standard deviations of the mixture componentes. Its length must be equal to the one of \code{pro} and \code{means}.}
}
\value{
\link{list} with a vector named \code{y} containing the sampled values and a vector named \code{clu} with the cluster membership of each data point.
}
\description{
This function generate a random sample from a mixture of uni-dimensional Gaussian distributed random variables.
}
