% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDFMM_export.R
\name{generate_data}
\alias{generate_data}
\title{generate_data}
\usage{
generate_data(
  d,
  K = 3,
  mu = c(-20, 0, 20),
  sd = c(1, 1, 1),
  n_j = rep(200, d),
  seed = 124123
)
}
\arguments{
\item{d}{\link{integer} the number of levels in the data.}

\item{K}{\link{integer} the total number of clusters.}

\item{mu}{\link{vector} of length \code{K} defining the means of the clusters.}

\item{sd}{\link{vector} of length \code{K} defining the standard deviations of the clusters.}

\item{n_j}{\link{vector} of length \code{d} defining the cardinality of each groups.}

\item{seed}{\link{integer} the random seed.}
}
\value{
\link{list} with a matrix of size \code{d x max(n_j)} named \code{data} containing the data to be fed to \code{\link{GDFMM_sampler}}
and a vector named \code{real_partition} with the cluster membership of each data point.
}
\description{
VECCHIA, NON è DA USARE
This function generate data to be used in the \code{\link{GDFMM_sampler}} or \code{\link{GDFMM_marginal_sampler}}.
A major drawback of this function is that the current formulation allows only to use equal mixture proportions.
This is indeed the old version, the newest function is \code{\link{generate_data_prob}}.
}
