% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDFMM_export.R
\name{input_handle}
\alias{input_handle}
\title{Handle input}
\usage{
input_handle(tb)
}
\arguments{
\item{tb}{\link{tibble} a tibble with three columns. First column contains the ID of the data,
second column contains the level membership of each data. Finally, the third column contains the numerical value of the variabile}
}
\description{
This function gets a tibble with three columns containing the data in long form. Data are handled and a list is returned.
}
