% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lastirling1}
\alias{lastirling1}
\title{Rigon - lastirling1}
\usage{
lastirling1(n)
}
\description{
Commento Ale: questa fuzione restituisce il vettore contenente il logaritmo dell'intero sviluppo dei moduli di un
numero di Stirling del primo tipo.
Per esempio, per n=4, ho |s(n,0)| = 0, |s(n,1)| = 6, |s(n,2)|= 11, |s(n,3)| = 6, |s(n,4)| = 4.
Quindi, lo sviluppo di 4 è un vettore di 4 elementi (lo 0 è banale, lo omettiamo)
}
