% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lastirlings1}
\alias{lastirlings1}
\title{Rigon - lastirlings1}
\usage{
lastirlings1(n)
}
\description{
Commento Ale, questa è l'estensione di lastirling1. Questa ritorna una matrice triangolare inferiori, l'argomento n
vuol dire che deve calcolare tutti gli sviluppi fino ad n, partendo da s(1) che è banalmente uguale ad uno. Anche
questa lavora in scala logaritmica.
}
