% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDFMM_export.R
\name{log_stable_sum}
\alias{log_stable_sum}
\title{log_stable_sum}
\usage{
log_stable_sum(a, is_log = T)
}
\arguments{
\item{a}{\link{vector} vector whose values must be summed.}

\item{is_log}{\link{bool} states if elements of a are already in log scale or not. If not, elements of a must be strictly positive.}
}
\description{
This functions computes log(sum_i(a_i)) using a stable formula for log values. Let us denote a* to the the maximum value of vector a which is attained when i = i*.
\eqn{log(sum_i(a_i)) = log(a*) + log[ 1 + sum_{i not i*}(exp{log(a_i) - log(a*)}) ]}
See that only the logarithm of the elements of a are needed. Hence, it is likely that one has already computed them in log scale. If so, set is_log = T
}
