% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{MCMC_conditional_c}
\alias{MCMC_conditional_c}
\title{MCMC_conditional_c}
\usage{
MCMC_conditional_c(
  data_list,
  n_iter,
  burn_in,
  thin,
  seed,
  P0_prior_name,
  FixPart,
  algorithm,
  option
)
}
\description{
MCMC_conditional_c
}
