% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{my_falling_factorial_old}
\alias{my_falling_factorial_old}
\title{Falling Factorial}
\usage{
my_falling_factorial_old(n, a)
}
\description{
\loadmathjax This function computes the falling factorial \mjseqn{ a_{n} }. See \code{\link{raising_factorial}} for details.
The raising (here denote with the upper apex) and the falling factorial (here denote with the lower apex) are related by the following relationship
\mjsdeqn{(a)_{n} = (-1)^{n}(a)^{n}}.
}
