% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{my_log_falling_factorial_old}
\alias{my_log_falling_factorial_old}
\title{log Falling Factorial}
\usage{
my_log_falling_factorial_old(n, a)
}
\description{
\loadmathjax This function computes the logarithm of the falling factorial \mjseqn{ a_{n} } using the gsl code for the log of Pochhammer symbol.
See \code{\link{my_falling_factorial}} and \code{\link{compute_Pochhammer}} for details.
}
