% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{my_logC}
\alias{my_logC}
\title{My logC}
\usage{
my_logC(n, scale, location)
}
\description{
This function is the recursive formula 2.69 in "Combinatorial methods in discrete distributions" book by Charalambides.
It returns an (n+1 x n+1) matrix containing all the log(|C(nn,k)|) numbers, for nn = 0,...,n+1 and k = 0,...,nn.
scale and location must be negative and non-positive, respectively.
As a consequence, it is memory expensive.
}
