% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{my_logC2}
\alias{my_logC2}
\title{Compute log of absolute values of non Central C number}
\usage{
my_logC2(n, scale, location)
}
\description{
\loadmathjax This is the main function in the computation of C numbers. It uses the (2.69) formula in the "Combinatorial methods in discrete distributions" book.
It computes \mjseqn{log(|C(n,k; scale, location)|)} for each k=0,...,n.
scale and location must be negative and non-positive, respectively.
It uses eigen objects, apparetly it is slower than using Rcpp vectors.
}
