% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{my_logC2_central}
\alias{my_logC2_central}
\title{My logC2 - central}
\usage{
my_logC2_central(n, scale)
}
\description{
This function is the recursive formula 2.69 in "Combinatorial methods in discrete distributions" book by Charalambides for central numbers.
This is the specialization of \code{\link{my_logC2}} for central C numbers.
scale must be negative.
}
