% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDFMM_export.R
\name{predictive_marginal}
\alias{predictive_marginal}
\title{compute the predictive distribution in a specific group when using a marginal sampler}
\usage{
predictive_marginal(idx_group, grid, fit, option)
}
\description{
compute the predictive distribution in a specific group when using a marginal sampler
}
