% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDFMM_export.R
\name{predictive_marginal_all_groups}
\alias{predictive_marginal_all_groups}
\title{predictive_marginal_all_groups}
\usage{
predictive_marginal_all_groups(grid, fit, option, burnin = 0)
}
\arguments{
\item{grid}{\link{vector} a grid where the normal kernel is evaluated.}

\item{fit}{\link{list} the output of \code{\link{GDFMM_sampler}}}

\item{burnin}{\link{integer} the number of draws from \code{\link{GDFMM_sampler}} that must be discarded.}
}
\value{
\link{list} of length \code{d} where each element is the return object of \code{\link{predictive_marginal}}.
}
\description{
This function computes the predictive distribution for all groups generated from the \code{\link{GDFMM_marginal_sampler}}.
}
