% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDFMM_export.R
\name{predictive_new_group}
\alias{predictive_new_group}
\title{predictive_new_group}
\usage{
predictive_new_group(grid, fit, burnin = 1, alpha_gamma, beta_gamma)
}
\arguments{
\item{grid}{\link{vector} a grid where the normal kernel is evaluated.}

\item{fit}{\link{list} the output of \code{\link{GDFMM_sampler}}}

\item{burnin}{\link{integer} the number of draws from \code{\link{GDFMM_sampler}} that must be discarded.}

\item{alpha_gamma}{\link{double} the shape parameter in the prior of gamma.}

\item{beta_gamma}{\link{double} the rate parameter in the prior of gamma.}
}
\value{
\link{matrix} of size \code{n x length(grid)} containing the quantiles of level \code{0.025,0.5,0.975}.
}
\description{
This function computes the predictive distribution for group \code{d+1} that has not been observed.
}
