% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDFMM_export.R
\name{predictive_players}
\alias{predictive_players}
\title{predictive_players}
\usage{
predictive_players(ID_ply, dt, fit, burnin = 1)
}
\arguments{
\item{ID_ply}{\link{integer} the index of the player of interest.}

\item{dt}{\link{list} This must be the return object of \code{\link{handle_input}}. The Clustering must be added as well}

\item{fit}{\link{list} the output of a conditional sampler, \code{\link{GDFMM_sampler}} or \code{\link{ConditionalSampler}}}

\item{burnin}{\link{integer} the number of draws from \code{\link{GDFMM_sampler}} that must be discarded.}
}
\value{
\link{matrix} of size ?? containing the quantiles of level \code{0.025,0.5,0.975}.
}
\description{
This function computes the predictive distribution for player \code{ID_ply} generated from \code{\link{ConditionalSampler}}
}
