% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDFMM_export.R
\name{rmix}
\alias{rmix}
\title{rmix : function to extract a random sample of dimension n from a Gaussian mixture
defined by the 3 vectors p, mu, sigma (that must have same length!)}
\usage{
rmix(n, p, mu, sigma)
}
\arguments{
\item{n}{dimension of the random sample that has to be extracted}

\item{p}{vector of weights for the components of the mixture}

\item{mu}{vector of means for the components}

\item{sigma}{vector of standard deviations for the components}
}
\value{
random sample derived my the specified mixture
}
\description{
rmix : function to extract a random sample of dimension n from a Gaussian mixture
defined by the 3 vectors p, mu, sigma (that must have same length!)
}
