% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDFMM_export.R
\name{set_options}
\alias{set_options}
\title{set_options}
\usage{
set_options(
  partition = NULL,
  Mstar0 = 2,
  nu = 1,
  Lambda0 = 3,
  mu0 = 0,
  sigma0 = 1,
  gamma0 = 1,
  Adapt_MH_hyp1 = 0.7,
  Adapt_MH_hyp2 = 0.234,
  Adapt_MH_power_lim = 10,
  Adapt_MH_var0 = 1,
  k0 = 1/10,
  nu0 = 10,
  alpha_gamma = 1,
  beta_gamma = 1,
  alpha_lambda = 1,
  beta_lambda = 1,
  UpdateU = T,
  UpdateM = T,
  UpdateGamma = T,
  UpdateS = T,
  UpdateTau = T,
  UpdateLambda = T
)
}
\arguments{
\item{UpdateLambda}{}
}
\description{
set_options
}
