% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDFMM_export.R
\name{simulate_data_OLD_POLIMI}
\alias{simulate_data_OLD_POLIMI}
\title{simulate_data (never checked)}
\usage{
simulate_data_OLD_POLIMI(
  n_simul,
  group_dim,
  p_mix,
  mu,
  sigma,
  burnin = 2000,
  n_iter = 2000,
  thin = 3,
  seed = 1234,
  option,
  dir = "."
)
}
\arguments{
\item{n_simul}{\link{integer} number of datasets to be simulated}

\item{group_dim}{\link{vector} of numerosity for each group in a dataset}

\item{p_mix}{\link{matrix} with each row that specifies weights for the mixture in the corrispondent
group}

\item{mu}{\link{vector} of means for the Gaussian components of the mixture}

\item{sigma}{\link{vector} of standard deviations for the Gaussian components of the mixture}

\item{burnin}{\link{integer} number of iterations to be discarded for the GDFMM}

\item{n_iter}{\link{integer} number of iterations to be saved for the GDFMM run}

\item{thin}{thinning of the GDFMM run}

\item{seed}{\link{integer} seed for the GDFMM run (0 ==> random seed)}

\item{option}{\link{list} of option for the GDFMM model. See GDFMM_sampler help for the list of
needed values}

\item{dir}{path to the directory where data about simlation have to be saved}
}
\value{
some metrics to evaluate the goodness of GDFMM
}
\description{
function to simulate data from the specified configuration of Gaussian
mixtures in different groups. Results for each one of the \code{n_simul}
datasets are saved in a directory create with the time-stamp of the simulation.
}
