% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDFMM_export.R
\name{sum_check}
\alias{sum_check}
\title{sum_check : check if elements in a vector sum to an expected value. Elements of the vector
are supposed to be computed as expected_sum*p. If expectations are not respected
return a corrected version of vec, according to p}
\usage{
sum_check(vec, expected_sum, p)
}
\arguments{
\item{vec}{vector to integer value to be checked}

\item{expected_sum}{expected value for sum(vec)}

\item{p}{theoretical weights applied to divide expected_sum in vec's elements}
}
\value{
corrected version of vec (if needed)
}
\description{
sum_check : check if elements in a vector sum to an expected value. Elements of the vector
are supposed to be computed as expected_sum*p. If expectations are not respected
return a corrected version of vec, according to p
}
