% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDFMM_export.R
\name{arrange_partition}
\alias{arrange_partition}
\title{arrange_partition:}
\usage{
arrange_partition(partition)
}
\arguments{
\item{partition}{\link{vector} with desidered partition}
}
\value{
\link{vector} containing the partition following the requirement of the sampler
}
\description{
DA FARE, SISTEMA LA PARTIZIONE PER COME LA VUOLE IL SAMPLER
}
