% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDFMM_export.R
\name{GDFMM_sampler}
\alias{GDFMM_sampler}
\title{GDFMM Gibbs Sampler: options for fixing the Partition (FixPartition) and for changing prior (P0.prior)}
\usage{
GDFMM_sampler(
  data,
  niter,
  burnin,
  thin,
  seed,
  P0.prior = "Normal-InvGamma",
  FixPartition = F,
  option
)
}
\arguments{
\item{data}{input data}

\item{niter}{number of iterations}

\item{burnin}{burnin period}

\item{thin}{thinning value}

\item{seed}{seed for GSL random engine}

\item{P0.prior}{string with the prior to be used as P0}

\item{FixPartition}{TRUE if we want to fix the partition}

\item{option}{list with initial values, hyperparameters and other options}
}
\value{
results of Gibbs Sampler
}
\description{
GDFMM Gibbs Sampler: options for fixing the Partition (FixPartition) and for changing prior (P0.prior)
}
