% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDFMM_export.R
\name{predictive_all_groups}
\alias{predictive_all_groups}
\title{pred_uninorm}
\usage{
predictive_all_groups(grid, fit, burnin = 1)
}
\arguments{
\item{grid}{\link{vector} a grid where the normal kernel is evaluated.}

\item{fit}{\link{list} the output of \code{\link{GDFMM_sampler}}}
}
\value{
\link{list} of length \code{d} where each element is the return object of \code{\link{predictive}}.
}
\description{
This function computes the predictive distribution for all groups generated from the \code{\link{GDFMM_sampler}}.
}
