% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDFMM_export.R
\name{set_options}
\alias{set_options}
\title{set_options}
\usage{
set_options(
  partition = NULL,
  Mstar0 = 2,
  Lambda0 = 3,
  mu0 = 0,
  sigma0 = 1,
  gamma0 = 1,
  Adapt_MH_hyp1 = 0.7,
  Adapt_MH_hyp2 = 0.234,
  Adapt_MH_power_lim = 10,
  Adapt_MH_var0 = 1,
  proposal_Mstar = 1,
  k0 = 1/10,
  nu0 = 10,
  alpha_gamma = 1,
  beta_gamma = 1,
  alpha_lambda = 1,
  beta_lambda = 1,
  init_mean_cluster = NULL,
  init_var_cluster = NULL,
  UpdateU = T,
  UpdateM = T,
  UpdateGamma = T,
  UpdateS = T,
  UpdateTau = T,
  UpdateLambda = T
)
}
\arguments{
\item{partition}{\link{vector} of length equal to the number of data. If \code{NULL}, all data points are put in the same cluster.}

\item{Mstar0}{\link{integer} the initial value of non-allocated components}

\item{Lambda0}{\link{double} the initial value for Lambda.}

\item{mu0}{\link{double} the mean parameter in the prior of mu.}

\item{sigma0}{\link{double} the rate parameter in the prior of sigma.}

\item{gamma0}{\link{double} the initial value for the gamma parameters.}

\item{Adapt_MH_hyp1}{\link{double} default is 0.7.}

\item{Adapt_MH_hyp2}{\link{double} default is 0.234.}

\item{Adapt_MH_power_lim}{\link{double} default is 10.}

\item{Adapt_MH_var0}{\link{double} default is 1.}

\item{proposal_Mstar}{\link{integer} must be strictly positive. The proposal distribution for Metropolis-Hasting update of Mstar is a discrete uniform
distribution between \code{-proposal_Mstar} and \code{proposal_Mstar}. 0 value is escluded.}

\item{k0}{\link{double} the parameter in the prior of mu.}

\item{nu0}{\link{double} the shape parameter in the prior of sigma.}

\item{alpha_gamma}{\link{double} the shape parameter in the prior of gamma.}

\item{beta_gamma}{\link{double} the rate parameter in the prior of gamma.}

\item{alpha_lambda}{\link{double} the shape parameter in the prior of lambda.}

\item{beta_lambda}{\link{double} the rate parameter in the prior of lambda.}

\item{UpdateU}{\link{bool} set \code{TRUE} if U must be updated. Set \code{FALSE} to fix it to a common value.}

\item{UpdateM}{\link{bool} set \code{TRUE} if Mstar must be updated. Set \code{FALSE} to fix it to a common value.}

\item{UpdateGamma}{\link{bool} set \code{TRUE} if gamma must be updated. Set \code{FALSE} to fix it to a common value.}

\item{UpdateS}{\link{bool} set \code{TRUE} if S must be updated. Set \code{FALSE} to fix it to a common value.}

\item{UpdateTau}{\link{bool} set \code{TRUE} if tau must be updated. Set \code{FALSE} to fix it to a common value.}

\item{UpdateLambda}{\link{bool} set \code{TRUE} if Lambda must be updated. Set \code{FALSE} to fix it to a common value.}
}
\description{
Use this function to set up the options for the conditional Gibbs sampler.
}
